Implement a table that behaves more like a figure.
For entering tabular data, it'd be nice for tables to behave more like
figures:

- The data should come from an external spreadsheet, rather than being
  embedded in the document itself.  The data is often more useful in an
  external spreadsheet.  Even if it isn't, entering tabular data in
  restructured text is extremely tedious.

- The caption should be paragraph-length.  The caption really needs to
  explain every column, and this just can't be done on a single line.

I adapted this directive from:

    https://github.com/hackerain/sphinxcontrib-excel-table

The table is rendered using "handsontable", which is basically a
javascript library for rendering Excel-like tables.  This allows you to
interact with the data directly in the notebook, which is pretty nice.
Obviously it won't work with LaTeX or anything, though.