Use pandas.read_excel() rather than openpyxl.
Openpyxl had a weird behavior where it would sometimes think that tables
had millions of rows or columns, and spend forever trying to parse them.
I'm pretty sure this was ultimately due to Gnumeric failing to
appropriately annotate the size of the spreadsheet, but there wasn't a
clear workaround.

Pandas doesn't seem to suffer from this problem, and it also supports
many more file formats.  It's a bigger dependency, but I think it's just
better overall.